Move the vector expression converters from core to their projects
