+++ vector-stores/spring-ai-pgvector-store/src/main/java/org/springframework/ai/vectorstore/PgVectorFilterExpressionConverter.java
@@ -13,26 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.ai.vectorstore.filter.converter;
+package org.springframework.ai.vectorstore;
 
 import org.springframework.ai.vectorstore.filter.Filter.Expression;
-import org.springframework.ai.vectorstore.filter.Filter.ExpressionType;
 import org.springframework.ai.vectorstore.filter.Filter.Group;
 import org.springframework.ai.vectorstore.filter.Filter.Key;
+import org.springframework.ai.vectorstore.filter.converter.AbstractFilterExpressionConverter;
 
 /**
- * Converts {@link Expression} into Milvus metadata filter expression format.
- * (https://milvus.io/docs/json_data_type.md)
+ * Converts {@link Expression} into PgVector metadata filter expression format.
+ * (https://www.postgresql.org/docs/current/functions-json.html)
  *
  * @author Christian Tzolov
  */
-public class MilvusFilterExpressionConverter extends AbstractFilterExpressionConverter {
+public class PgVectorFilterExpressionConverter extends AbstractFilterExpressionConverter {
 
 	@Override
-	protected void doExpression(Expression exp, StringBuilder context) {
-		this.convertOperand(exp.left(), context);
-		context.append(getOperationSymbol(exp));
-		this.convertOperand(exp.right(), context);
+	protected void doExpression(Expression expression, StringBuilder context) {
+		this.convertOperand(expression.left(), context);
+		context.append(getOperationSymbol(expression));
+		this.convertOperand(expression.right(), context);
 	}
 
 	private String getOperationSymbol(Expression exp) {
@@ -58,19 +58,23 @@ public class MilvusFilterExpressionConverter extends AbstractFilterExpressionCon
 			case NIN:
 				return " nin ";
 			default:
-				throw new RuntimeException("Not supported expression type:" + exp.type());
+				throw new RuntimeException("Not supported expression type: " + exp.type());
 		}
 	}
 
 	@Override
-	protected void doGroup(Group group, StringBuilder context) {
-		this.convertOperand(new Expression(ExpressionType.AND, group.content(), group.content()), context); // trick
+	protected void doKey(Key key, StringBuilder context) {
+		context.append("$." + key.key());
 	}
 
 	@Override
-	protected void doKey(Key key, StringBuilder context) {
-		var identifier = (hasOuterQuotes(key.key())) ? removeOuterQuotes(key.key()) : key.key();
-		context.append("metadata[\"" + identifier + "\"]");
+	protected void doStartGroup(Group group, StringBuilder context) {
+		context.append("(");
+	}
+
+	@Override
+	protected void doEndGroup(Group group, StringBuilder context) {
+		context.append(")");
 	}
 
 }
